test: move inline test modules to sibling files
Production files carried inline #[cfg(test)] mod tests { ... } blocks. Extract
each to a sibling tests.rs (lib->src/tests.rs, error/normalize/parser/validate
->src/<mod>/tests.rs) declared with #[cfg(test)] mod tests;, so production
sources hold only production code. No test logic changed (135 tests, same set).
